% Auto-generated: do not edit by hand
\name{stripeDash}

\alias{stripeDash}

\title{StripeDash component}

\description{

}

\usage{
stripeDash(id=NULL, amount=NULL, customConfirmMessage=NULL,
errorMessage=NULL, label=NULL, paymentIntentId=NULL,
paymentMethodDetails=NULL, paymentMethodId=NULL,
paymentStatus=NULL, prePaymentMessage=NULL,
referenceId=NULL, stripe_api=NULL, stripe_key=NULL)
}

\arguments{
\item{id}{Character. }

\item{amount}{Numeric. }

\item{customConfirmMessage}{Character. }

\item{errorMessage}{Character. }

\item{label}{Character. }

\item{paymentIntentId}{Character. }

\item{paymentMethodDetails}{Character. }

\item{paymentMethodId}{Character. }

\item{paymentStatus}{Character. }

\item{prePaymentMessage}{Character. }

\item{referenceId}{Character. }

\item{stripe_api}{Character. }

\item{stripe_key}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

